requirements.txt
numpy
pandas
matplotlib
seaborn
scikit-learn
scipy
pytorch
accelerate
nltk
sae_lens
datasets
huggingface_hub
transformer_lens
sae_dashboard
python-dotenv